refactor(api): move findPaginatedfilteredScoParticipant into prescription bounded context
