+++ api/lib/domain/usecases/index.js
@@ -143,7 +143,6 @@ import * as resetPasswordDemandRepository from '../../infrastructure/repositorie
 import * as resetPasswordService from '../../domain/services/reset-password-service.js';
 import * as scoAccountRecoveryService from '../services/sco-account-recovery-service.js';
 import * as scoCertificationCandidateRepository from '../../infrastructure/repositories/sco-certification-candidate-repository.js';
-import * as scoOrganizationParticipantRepository from '../../infrastructure/repositories/sco-organization-participant-repository.js';
 import * as scorecardService from '../../domain/services/scorecard-service.js';
 import * as scoringCertificationService from '../../domain/services/scoring/scoring-certification-service.js';
 import * as sessionCodeService from '../../../src/certification/session/domain/services/session-code-service.js';
@@ -356,7 +355,6 @@ const dependencies = {
   schoolRepository,
   scoAccountRecoveryService,
   scoCertificationCandidateRepository,
-  scoOrganizationParticipantRepository,
   scorecardService,
   scoringCertificationService,
   sessionCodeService,
