+++ api/src/prescription/learner-list/domain/read-models/ScoOrganizationParticipant.js
