+++ api/src/prescription/learner-list/routes.js
@@ -1,5 +1,6 @@
 import * as supLearnerListRoute from './application/sup-learner-list-route.js';
+import * as scoLearnerListRoute from './application/sco-learner-list-route.js';
 
-const learnerListRoutes = [supLearnerListRoute];
+const learnerListRoutes = [supLearnerListRoute, scoLearnerListRoute];
 
 export { learnerListRoutes };
