+++ api/tests/integration/application/organizations/organization-controller_test.js
@@ -3,7 +3,6 @@ import { securityPreHandlers } from '../../../../lib/application/security-pre-ha
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 
 import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
-import { ScoOrganizationParticipant } from '../../../../lib/domain/read-models/ScoOrganizationParticipant.js';
 import { getCertificationAttestationsPdf as certificationAttestationPdf } from '../../../../lib/infrastructure/utils/pdf/certification-attestation-pdf.js';
 import * as moduleUnderTest from '../../../../lib/application/organizations/index.js';
 import { NoCertificationAttestationForDivisionError } from '../../../../lib/domain/errors.js';
@@ -15,7 +14,6 @@ describe('Integration | Application | Organizations | organization-controller', 
   beforeEach(async function () {
     sandbox = sinon.createSandbox();
     sandbox.stub(usecases, 'findPaginatedFilteredOrganizationMemberships');
-    sandbox.stub(usecases, 'findPaginatedFilteredScoParticipants');
     sandbox.stub(usecases, 'createOrganizationInvitations');
     sandbox.stub(usecases, 'acceptOrganizationInvitation');
     sandbox.stub(usecases, 'findPendingOrganizationInvitations');
@@ -120,53 +118,6 @@ describe('Integration | Application | Organizations | organization-controller', 
         expect(response.result.included[0].id).to.equal(`${membership.organization.id}`);
         expect(response.result.included[1].type).to.equal('users');
         expect(response.result.included[1].id).to.equal(`${membership.user.id}`);
-      });
-    });
-  });
-
-  describe('#findPaginatedFilteredScoParticipants', function () {
-    context('Success cases', function () {
-      it('should return an HTTP response with status code 200', async function () {
-        // given
-        const scoOrganizationParticipant = new ScoOrganizationParticipant();
-        usecases.findPaginatedFilteredScoParticipants.resolves({ data: [scoOrganizationParticipant] });
-        securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.returns(true);
-
-        // when
-        const response = await httpTestServer.request('GET', '/api/organizations/1234/sco-participants');
-
-        // then
-        expect(response.statusCode).to.equal(200);
-      });
-
-      it('should return an HTTP response formatted as JSON:API', async function () {
-        // given
-        const scoOrganizationParticipant = new ScoOrganizationParticipant();
-        usecases.findPaginatedFilteredScoParticipants.resolves({ data: [scoOrganizationParticipant] });
-        securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.returns(true);
-
-        // when
-        const response = await httpTestServer.request('GET', '/api/organizations/1234/sco-participants');
-
-        // then
-        expect(response.result.data[0].type).to.equal('sco-organization-participants');
-      });
-    });
-
-    context('Error cases', function () {
-      context('when user is not allowed to access resource', function () {
-        it('should resolve a 403 HTTP response', async function () {
-          // given
-          securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.callsFake((request, h) => {
-            return Promise.resolve(h.response().code(403).takeover());
-          });
-
-          // when
-          const response = await httpTestServer.request('GET', '/api/organizations/1234/sco-participants');
-
-          // then
-          expect(response.statusCode).to.equal(403);
-        });
       });
     });
   });
