+++ api/tests/prescription/learner-list/unit/domain/usecases/find-paginated-filtered-sco-participants_test.js
@@ -1,5 +1,5 @@
-import { expect, sinon } from '../../../test-helper.js';
-import { findPaginatedFilteredScoParticipants } from '../../../../lib/domain/usecases/find-paginated-filtered-sco-participants.js';
+import { expect, sinon } from '../../../../../test-helper.js';
+import { findPaginatedFilteredScoParticipants } from '../../../../../../src/prescription/learner-list/domain/usescases/find-paginated-filtered-sco-participants.js';
 
 describe('Unit | UseCase | findPaginatedFilteredScoParticipants', function () {
   let scoOrganizationParticipantRepository;
