+++ api/tests/unit/application/organizations/organization-controller_test.js
@@ -3,7 +3,6 @@ import { domainBuilder, expect, generateValidRequestAuthorizationHeader, hFake, 
 import { Organization } from '../../../../lib/domain/models/Organization.js';
 import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
 import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { ScoOrganizationParticipant } from '../../../../lib/domain/read-models/ScoOrganizationParticipant.js';
 import { organizationController } from '../../../../lib/application/organizations/organization-controller.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import * as queryParamsUtils from '../../../../lib/infrastructure/utils/query-params-utils.js';
@@ -541,169 +540,6 @@ describe('Unit | Application | Organizations | organization-controller', functio
             },
           },
         ],
-      });
-    });
-  });
-
-  describe('#findPaginatedFilteredScoParticipants', function () {
-    const connectedUserId = 1;
-    const organizationId = 145;
-
-    let scoOrganizationParticipant;
-    let serializedScoOrganizationParticipant;
-    let request;
-    let dependencies;
-
-    beforeEach(function () {
-      request = {
-        auth: { credentials: { userId: connectedUserId } },
-        params: { id: organizationId },
-      };
-
-      sinon.stub(usecases, 'findPaginatedFilteredScoParticipants');
-
-      const scoOrganizationParticipantsSerializerStub = {
-        serialize: sinon.stub(),
-      };
-
-      dependencies = {
-        queryParamsUtils,
-        scoOrganizationParticipantsSerializer: scoOrganizationParticipantsSerializerStub,
-      };
-
-      scoOrganizationParticipant = new ScoOrganizationParticipant();
-      serializedScoOrganizationParticipant = {
-        data: [
-          {
-            ...scoOrganizationParticipant,
-            isAuthenticatedFromGAR: false,
-          },
-        ],
-        meta: { pagination: { page: 1 } },
-      };
-    });
-
-    it('should call the usecase to find sco participants with users infos related to the organization id', async function () {
-      // given
-      usecases.findPaginatedFilteredScoParticipants.resolves({});
-
-      // when
-      await organizationController.findPaginatedFilteredScoParticipants(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredScoParticipants).to.have.been.calledWithExactly({
-        organizationId,
-        filter: {},
-        page: {},
-        sort: {},
-      });
-    });
-
-    it('should call the usecase to find sco participants with users infos related to filters', async function () {
-      // given
-      request = {
-        ...request,
-        query: {
-          'filter[lastName]': 'Bob',
-          'filter[firstName]': 'Tom',
-          'filter[connectionTypes]': 'email',
-          'filter[divisions][]': 'D1',
-        },
-      };
-      usecases.findPaginatedFilteredScoParticipants.resolves({});
-
-      // when
-      await organizationController.findPaginatedFilteredScoParticipants(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredScoParticipants).to.have.been.calledWithExactly({
-        organizationId,
-        filter: { lastName: 'Bob', firstName: 'Tom', connectionTypes: ['email'], divisions: ['D1'] },
-        page: {},
-        sort: {},
-      });
-    });
-    it('should call the usecase to find sco participants with users infos related to sort', async function () {
-      // given
-      request = {
-        ...request,
-        query: {
-          'sort[participationCount]': 'asc',
-          'sort[lastnameSort]': 'asc',
-          'sort[divisionSort]': 'desc',
-        },
-      };
-      usecases.findPaginatedFilteredScoParticipants.resolves({});
-
-      // when
-      await organizationController.findPaginatedFilteredScoParticipants(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredScoParticipants).to.have.been.calledWithExactly({
-        organizationId,
-        filter: {},
-        page: {},
-        sort: {
-          participationCount: 'asc',
-          lastnameSort: 'asc',
-          divisionSort: 'desc',
-        },
-      });
-    });
-
-    it('should call the usecase to find sco participants with users infos related to pagination', async function () {
-      // given
-      request = { ...request, query: { 'page[size]': 10, 'page[number]': 1 } };
-      usecases.findPaginatedFilteredScoParticipants.resolves({});
-
-      // when
-      await organizationController.findPaginatedFilteredScoParticipants(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredScoParticipants).to.have.been.calledWithExactly({
-        organizationId,
-        filter: {},
-        page: { size: 10, number: 1 },
-        sort: {},
-      });
-    });
-
-    it('should return the serialized sco participants belonging to the organization', async function () {
-      // given
-      const meta = { pagination: { page: 1 } };
-      const scoOrganizationParticipants = [scoOrganizationParticipant];
-      usecases.findPaginatedFilteredScoParticipants.resolves({ data: scoOrganizationParticipants, meta });
-      dependencies.scoOrganizationParticipantsSerializer.serialize
-        .withArgs({ scoOrganizationParticipants, meta })
-        .returns(serializedScoOrganizationParticipant);
-
-      // when
-      const response = await organizationController.findPaginatedFilteredScoParticipants(request, hFake, dependencies);
-      // then
-      expect(response).to.deep.equal(serializedScoOrganizationParticipant);
-    });
-
-    it('map all certificability values', async function () {
-      // given
-      request = {
-        auth: { credentials: { userId: connectedUserId } },
-        params: { id: organizationId },
-        query: {
-          'filter[certificability][]': ['eligible', 'non-eligible', 'not-available'],
-        },
-      };
-      usecases.findPaginatedFilteredScoParticipants.resolves({ data: [] });
-      dependencies.scoOrganizationParticipantsSerializer.serialize.returns({});
-
-      // when
-      await organizationController.findPaginatedFilteredScoParticipants(request, hFake);
-
-      // then
-      expect(usecases.findPaginatedFilteredScoParticipants).to.have.been.calledWithExactly({
-        organizationId,
-        filter: { certificability: [true, false, null] },
-        page: {},
-        sort: {},
       });
     });
   });
