+++ api/tests/unit/infrastructure/serializers/jsonapi/organization/sco-organization-participants-serializer_test.js
@@ -1,5 +1,5 @@
 import { expect } from '../../../../../test-helper.js';
-import { ScoOrganizationParticipant } from '../../../../../../lib/domain/read-models/ScoOrganizationParticipant.js';
+import { ScoOrganizationParticipant } from '../../../../../../src/prescription/learner-list/domain/read-models/ScoOrganizationParticipant.js';
 import * as serializer from '../../../../../../lib/infrastructure/serializers/jsonapi/organization/sco-organization-participants-serializer.js';
 import { CampaignParticipationStatuses as campaignParticipationsStatuses } from '../../../../../../lib/domain/models/CampaignParticipationStatuses.js';
